Restore durability as a float rather than an int.

+++ app/src/main/java/com/watabou/pixeldungeon/items/Item.java
@@ -541,7 +541,7 @@ public class Item implements Bundlable {
 		cursed	= bundle.getBoolean( CURSED );
 		
 		if (isUpgradable()) {
-			durability = bundle.getInt( DURABILITY );
+			durability = bundle.getFloat( DURABILITY );
 		}
 		
 		QuickSlot.restore( bundle, this );
